refactor(cli): simplify coordinator post-1.4.0 code quality
- Extract magic timeout numbers into named constants
  (EXPAND_LOCK_TIMEOUT, MERGE_LOCK_TIMEOUT)
- Skip no-op DAG writes in _dagTransaction via snapshot comparison
- Remove provenance-only comments, keep structural WHY explanations
- Extract shared test helpers into coordinator-test-helpers.js,
  eliminating ~80 lines of duplication across 3 race test files
- Fix Biome format in normalizeStatus error throw
